# money-core/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is API Documentation of Money Thrower Core",
        "title": "Money Thrower Core API",
        "contact": {
            "name": "Trung Dung Vu",
            "email": "trungdungvu172@gmail.com"
        },
        "version": "1.0"
    },
    "host": "api.moneythrower.site",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login with email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Login with email and password",
                "parameters": [
                    {
                        "description": "Login with account",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/view.LoginForm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Logout and push JWT Token into blacklist for 72 hours",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Logout and push JWT Token into blacklist for 72 hours",
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register with email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Register with email and password",
                "parameters": [
                    {
                        "description": "Register with account",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/view.RegisterForm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    }
                }
            }
        },
        "/dummy": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Dummy test connection and token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "testing"
                ],
                "summary": "Dummy test connection and token",
                "responses": {
                    "200": {
                        "description": "message\": \"Hello World\"}",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/password/change": {
            "post": {
                "description": "Submit token, new password, email to reset password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "password management"
                ],
                "summary": "Submit token, new password, email to reset password",
                "parameters": [
                    {
                        "description": "Submit token, new password, email to reset password",
                        "name": "ForgotPassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.SubmitNewPasswordForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message\": true}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    }
                }
            }
        },
        "/password/forgot": {
            "post": {
                "description": "Submit email for sending reset password email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "password management"
                ],
                "summary": "Submit email for sending reset password email",
                "parameters": [
                    {
                        "description": "Submit email for reset password (Must be valid email)",
                        "name": "ForgotPassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.ForgotPasswordForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message\": true}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    }
                }
            }
        },
        "/wallet": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Edit info of a wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Edit info of a wallet",
                "parameters": [
                    {
                        "description": "Update wallet",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.WalletForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/view.WalletForm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Add a new wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Add a new wallet",
                "parameters": [
                    {
                        "description": "Create wallet",
                        "name": "create",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.WalletForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/view.WalletForm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.AppError": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "null"
                    ]
                },
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "invalid form"
                }
            }
        },
        "view.ForgotPasswordForm": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@gmail.com"
                }
            }
        },
        "view.LoginForm": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "test"
                }
            }
        },
        "view.RegisterForm": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@gmail.com"
                },
                "id": {
                    "type": "string",
                    "example": "id"
                },
                "password": {
                    "type": "string",
                    "example": "test"
                },
                "token": {
                    "type": "string",
                    "example": "token-string"
                }
            }
        },
        "view.SubmitNewPasswordForm": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "test"
                },
                "token": {
                    "type": "string",
                    "example": "token-string"
                }
            }
        },
        "view.WalletForm": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number",
                    "example": 1500000
                },
                "currency": {
                    "type": "string",
                    "example": "VND"
                },
                "icon": {
                    "type": "string",
                    "example": "mua_sam"
                },
                "id": {
                    "type": "string",
                    "example": "id"
                },
                "name": {
                    "type": "string",
                    "example": "MyBankAccount"
                },
                "type": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}